can increment by 1Hz in GUI. Can set freq in terms of sysclk period (60hz res). Need to figure out how to set in terms of hr period as well (1hz res).

+++ gui.h
@@ -15,10 +15,10 @@
 extern uint16_t receivedChar;
 extern unsigned char *msg;
 
-extern int dutyCycle;
+extern uint16_t dutyCycle;
 extern double dutyCycleTrack;
-extern unsigned int period;
-extern int guiState;
+extern uint16_t period;
+extern uint16_t guiState;
 
 // functions
 void run_main_menu();
